hope-project/backend/requirements.txt
# ============================================================
# HOPE — Backend Python Dependencies
# Save to: backend/requirements.txt
#
# Install: pip install -r requirements.txt
# ============================================================

# ── Web Framework ──
Flask==3.0.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0

# ── Database ──
pymongo==4.6.2
dnspython==2.6.1          # required for mongodb+srv:// URIs

# ── AI / ML ──
tensorflow==2.16.1
numpy==1.26.4
Pillow==10.2.0

# ── Cloud Storage ──
cloudinary==1.39.1

# ── Security ──
bcrypt==4.1.2
python-dotenv==1.0.1

# ── Utilities ──
python-dateutil==2.9.0

# ── Email ──
Flask-Mail==0.10.0

# ── Testing ──
pytest==8.1.1
pytest-flask==1.3.0
